+++ docutils/docutils/__init__.py
@@ -57,9 +57,8 @@ from collections import namedtuple
 TYPE_CHECKING = False
 if TYPE_CHECKING:
     from collections.abc import Sequence
-    from typing import Any, ClassVar, Literal, Protocol, Union
+    from typing import Any, ClassVar, Literal, Union
 
-    from docutils.nodes import Element
     from docutils.transforms import Transform
 
     _Components = Literal['reader', 'parser', 'writer', 'input', 'output']
@@ -74,14 +73,6 @@ if TYPE_CHECKING:
               str|None, str|None, Sequence[_OptionTuple]],
         ]
 
-    class _UnknownReferenceResolver(Protocol):
-        """Deprecated. Will be removed in Docutils 1.0."""
-        # See `TransformSpec.unknown_reference_resolvers`.
-
-        priority: int
-
-        def __call__(self, node: Element, /) -> bool:
-            ...
 
 __docformat__ = 'reStructuredText'
 
@@ -281,43 +272,6 @@ class TransformSpec:
 
     # Deprecated; for compatibility.
     default_transforms: ClassVar[tuple[()]] = ()
-
-    unknown_reference_resolvers: Sequence[_UnknownReferenceResolver] = ()
-    """List of hook functions which assist in resolving references.
-
-    Deprecated. Will be removed in Docutils 1.0
-    """
-    # Override in subclasses to implement component-specific resolving of
-    # unknown references.
-    #
-    # Unknown references have a 'refname' attribute which doesn't correspond
-    # to any target in the document.  Called when the transforms in
-    # `docutils.transforms.references` are unable to find a correct target.
-    #
-    # The list should contain functions which will try to resolve unknown
-    # references, with the following signature::
-    #
-    #     def reference_resolver(node: nodes.Element) -> bool:
-    #         '''Returns boolean: true if resolved, false if not.'''
-    #
-    # If the function is able to resolve the reference, it should also remove
-    # the 'refname' attribute and mark the node as resolved::
-    #
-    #     del node['refname']
-    #     node.resolved = True
-    #
-    # Each function must have a "priority" attribute which will affect the
-    # order the unknown_reference_resolvers are run
-    # cf. ../docs/api/transforms.html#transform-priority-range-categories ::
-    #
-    #     reference_resolver.priority = 500
-    #
-    # Examples:
-    #   The `MoinMoin ReStructured Text Parser`__ provided a resolver for
-    #   "WikiWiki links" in the 1.9 version.
-    #
-    #   __ https://github.com/moinwiki/moin-1.9/blob/1.9.11/MoinMoin/parser/
-    #      text_rst.py
 
 
 class Component(SettingsSpec, TransformSpec):